+++ services/ratings_service.py
@@ -1,6 +1,8 @@
 from analysis import read_custom_ratings
-from xlsx_to_json import is_excel_valid, prepare_json
-from utils.utils_endpoints import return_file_path
+from xlsx_to_json import is_excel_valid, upload_json
+
+from services.storage_service import StorageService
+from utils.json_prep import clean_ratings_data
 
 from typing import Dict, Any
 import json
@@ -8,10 +10,13 @@ import pandas as pd
 
 class RatingsService:
 
-    @staticmethod
-    async def process_uploaded_ratings(contents_of_xlsx: bytes, filename: str) -> str:
-        # Move the upload logic here
-        xlsx_ratings_sheet = pd.read_excel(contents_of_xlsx, # type: ignore
+    def __init__(self, contents:bytes, filename:str):
+        self.storage_service = StorageService(filename)
+        self.contents = contents
+
+
+    async def process_ratings(self) -> dict:
+        xlsx_ratings_sheet = pd.read_excel(self.contents, # type: ignore
                                            sheet_name=2,
                                            header=2,
                                            index_col=0,
@@ -20,7 +25,10 @@ class RatingsService:
         if not is_excel_valid(xlsx_ratings_sheet):
             raise ValueError("Invalid Excel format")
 
-        return prepare_json(xlsx_ratings_sheet, filename)
+        return clean_ratings_data(xlsx_ratings_sheet, self.storage_service.extract_date())
+
+    async def upload_ratings(self, df_as_dict):
+        return upload_json(df_as_dict, self.storage_service.storage_path())
 
     @staticmethod
     async def get_ratings(file_path: str, time_range: tuple[str, str], channels: list[str]) -> Dict[str, Any]: