+++ utils/json_prep.py
@@ -1,41 +1,16 @@
 ﻿import pandas as pd
 
-from utils.config import current_config
-from pathlib import Path
+from services.storage_service import StorageService
 from utils.logger import get_logger
 from typing import Final
-from datetime import datetime
 
 logger = get_logger(__name__)
 
 INDEX_COLUMN: Final = 'Timebands'
 
 
-def generate_storage_path(original_filename:str) -> str:
-    year, month, day = extract_date_from_filename(original_filename)
-    return f"{year}/{month}/{year}-{month}-{day}.json"
-
-
-def handle_storage_path(base_path: str) -> str:
-    """Handle storage-specific path operations
-
-    Example:
-        >>> received_path = ...
-
-    """
-    if current_config.STORAGE_TYPE == 'local':
-        full_path: Path = Path(f"{current_config.STORAGE_PATH}/{base_path}")
-        if full_path.exists():
-            logger.info(f"File {full_path} already exists, skipping...")
-            return str(full_path)
-        full_path.parent.mkdir(parents=True, exist_ok=True)
-        return str(full_path)
-    else:  # firebase
-        return f"{current_config.STORAGE_PATH}/{base_path}"
-
-
-def clean_ratings_data(dataframe, original_filename) -> dict:
-    year, month, day = extract_date_from_filename(original_filename)
+def clean_ratings_data(dataframe, date:tuple[str, str, str]) -> dict:
+    year, month, day = date
 
     file_format_from_date: str = f"{year}-{month}-{day}"
 
@@ -62,16 +37,6 @@ def clean_ratings_data(dataframe, original_filename) -> dict:
     }
     return json_data
 
-
-def extract_date_from_filename(filename: str) -> tuple[str, str, str]:
-    """Extract year, month, day from the filename in YYYY-MM-DD format."""
-    date_str = str(filename).split(' ')[-1].replace('.xlsx', '')
-    date_obj = datetime.strptime(date_str, '%Y-%m-%d')
-    return (str(date_obj.year),
-            f"{date_obj.month:02d}",
-            f"{date_obj.day:02d}")
-
-
 def fix_broadcast_time(timestring: str, date_of_file: str) -> str | None:
     """Convert time strings like '24:00' to next day datetime."""
     try:
@@ -83,8 +48,3 @@ def fix_broadcast_time(timestring: str, date_of_file: str) -> str | None:
         return f'{date_of_file} {hour:02d}:{minute:02d}'
     except:
         return None
-
-
-def rename_columns(df, string_to_remove):
-    new_columns: dict = {col: col.replace(string_to_remove, '') for col in df.columns}
-    return df.rename(columns=new_columns)  ## Not used!