+++ changelog.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on Keep a Changelog,
 and this project adheres to Semantic Versioning.
 
-[0.3.1] - 2025-06-08 
+[0.3.2] - 2025-06-13 
 -
+- Added StorageService class
+- Preparing database migration to PostgreSQL (Supabase)
+
+[0.3.1] - 2025-06-08
 - Latest .xlsx upload auto-reads the json file with default times (20:00–22:59)
 - Refactored API logic